requirements.txt
matplotlib==3.1.1
pandas==0.25.1
requests==2.22.0
playsound==1.2.2
Werkzeug==0.16.0
tensorflow==2.3.0
Flask==1.1.1
streamlit==0.70.0
numpy==1.16.5
gevent==1.4.0
gTTS==2.2.1
Keras==2.4.3
numpy==1.20.1
Pillow==8.1.2
